Refactor upload functionalities out of views
